requirements.txt
django==3.1.1
mysqlclient==1.4.6
bleach==3.2.0
pillow==7.2.0
martor==1.5.6
gspread==3.6.0
oauth2client==4.1.3
requests==2.24.0
djangorestframework==3.12.1